requirements-dev.txt
-r requirements.txt
pre-commit
detect-secrets
pytest
pytest-django
pytest-cov
factory-boy
selenium
webdriver-manager
